requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Configuration and Utilities
pyyaml>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
plotly>=5.15.0
tensorboard>=2.13.0

# Transformers and Multimodal Models
transformers>=4.30.0
sentencepiece>=0.1.99
accelerate>=0.20.0
open_clip_torch>=2.20.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# NLP and Text Processing
spacy>=3.6.0
nltk>=3.8.0
sacrebleu>=2.3.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Evaluation and Metrics
jiwer>=3.0.0
faiss-cpu>=1.7.4
hnswlib>=0.7.0

# Web Framework and Demo
streamlit>=1.25.0
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Development and Testing
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Tracking and Monitoring
wandb>=0.15.0
mlflow>=2.5.0
